Split NSPersistentContainer extension
