+++ Sources/UtilityBeltData/CoreData/NSPersistentContainer+Mocked.swift
@@ -3,29 +3,6 @@
 import CoreData
 
 public extension NSPersistentContainer {
-    /// The persistent container store type.
-    ///
-    /// Does **NOT** include the XML store type, which is only available on macOS 10.4+.
-    enum StoreType {
-        /// Stores data in binary.
-        case binary
-        /// Stores data in-memory.
-        case memory
-        /// Stores data in a SQLite database.
-        case sqlite
-
-        var rawValue: String {
-            switch self {
-            case .binary:
-                return NSBinaryStoreType
-            case .memory:
-                return NSInMemoryStoreType
-            case .sqlite:
-                return NSSQLiteStoreType
-            }
-        }
-    }
-
     /// Creates a mocked persistent container that stores data in memory instead of on disk.
     /// - Parameter name: The name of the Data Model to persist.
     static func mocked(name: String, storeType: StoreType = .memory) -> NSPersistentContainer {
